Created MongoQueries.Java to generalize mongodb queries even more!
