+++ v0ncentartsite/server/src/main/java/com/v0ncent/server/POJOMODELS/Model.java
@@ -1,3 +1,10 @@
+/*
+ * --------------------------------------------------------------------------
+ *                   Server Application Model Parent Class
+ * This class is responsible for parenting all POJO Model objects
+ * @author v0ncent
+ * --------------------------------------------------------------------------
+ */
 package com.v0ncent.server.POJOMODELS;
 public class Model {
     public Model(){
