+++ v0ncentartsite/server/src/main/java/com/v0ncent/server/RestAPIS/GalleryController.java
@@ -24,7 +24,7 @@ import java.util.Collection;
 import java.util.Optional;
 @RestController
 @RequestMapping("/api")
-public class GalleryController {
+public final class GalleryController {
     private static final Logger LOGGER = LoggerFactory.getLogger(GalleryController.class);
     @Autowired //<--- field inject these nuts lmfao. No but for real I rly don't care for this warning, but it won't let me ignore it >:d
     private GalleryTool galleryTool;
